perf(rust, python): use a trusted anyvalue builder
